Use f64 instead of f32 in yat-rack
It only needs to be f32 for the output sample, so can be converted at
this point (basically a truncated amplitude with more than enough
resolution).

In the actual audio processing phase, particularly for clock
incrementation, the increased resolution could be better for quality, at
no extra cost (in theory).

+++ yat-rack/src/modules/oscillator.rs
@@ -59,7 +59,7 @@ impl PartialEq for Oscillator {
 impl IoModule for Oscillator {
     /// Read inputs and populate outputs
     fn process_inputs(&mut self) {
-        let pi = std::f32::consts::PI;
+        let pi = std::f64::consts::PI;
 
         let time = self
             .clock
